Update pcitd server and agent
