+++ app/Console/PCITDaemon/Agent.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Console\PCITDaemon;
 
+use App\Build;
 use App\Console\Events\LogHandle;
 use App\Console\Events\Subject;
 use App\Console\Events\UpdateBuildStatus;
@@ -13,19 +14,14 @@ use PCIT\Support\CI;
 use PCIT\Support\Log;
 
 /**
- * Class Agent.
- *
- * Agent run job
+ * Agent run job, need docker.
  */
-class Agent
+class Agent extends Kernel
 {
     /**
-     * @param int $build_key_id
-     * @param     $config
-     *
      * @throws \Exception
      */
-    public function handle(int $build_key_id, $config): void
+    public function handle(): void
     {
         Log::debug(__FILE__, __LINE__, 'Docker connect ...');
 
@@ -33,34 +29,36 @@ class Agent
 
         Log::debug(__FILE__, __LINE__, 'Docker build Start ...');
 
-        $job_ids = Job::getByBuildKeyID($build_key_id, true);
+        // 取出一个 job,包括 job config, build key id
+        $job_data = Job::getPendingJob()[0];
+
+        ['id' => $job_id, 'build_id' => $build_key_id] = $job_data;
+
+        $config = Build::getConfig((int) $build_key_id);
 
         $subject = new Subject();
 
-        foreach ($job_ids as $job_id) {
-            $job_id = $job_id['id'];
-            Log::debug(__FILE__, __LINE__, 'Handle build jobs', ['job_id' => $job_id], Log::EMERGENCY);
+        Log::debug(__FILE__, __LINE__, 'Handle build jobs', ['job_id' => $job_id], Log::EMERGENCY);
+
+        $subject
+            // update build status in progress
+            ->register(new UpdateBuildStatus((int) $job_id, $config, CI::GITHUB_CHECK_SUITE_STATUS_IN_PROGRESS))
+            ->handle();
+
+        try {
+            (new PCIT())->build_agent->handle((int) $job_id);
+        } catch (\Throwable $e) {
+            Log::debug(__FILE__, __LINE__, 'Handle job success', ['job_id' => $job_id, 'message' => $e->getMessage()], Log::EMERGENCY);
 
             $subject
-                // update build status in progress
-                ->register(new UpdateBuildStatus((int) $job_id, $config, CI::GITHUB_CHECK_SUITE_STATUS_IN_PROGRESS))
+                ->register(new LogHandle((int) $job_id))
+                ->register(new UpdateBuildStatus((int) $job_id, $config, $e->getMessage()))
                 ->handle();
-
-            try {
-                (new PCIT())->build_agent->handle((int) $job_id);
-            } catch (\Throwable $e) {
-                Log::debug(__FILE__, __LINE__, 'Handle job success', ['job_id' => $job_id, 'message' => $e->getMessage()], Log::EMERGENCY);
-
-                $subject
-                    ->register(new LogHandle((int) $job_id))
-                    ->register(new UpdateBuildStatus((int) $job_id, $config, $e->getMessage()))
-                    ->handle();
-            }
         }
 
-        // update build status
-        $status = Job::getBuildStatusByBuildKeyId($build_key_id);
+        // 运行一个 job 之后更新 build 状态
+        $status = Job::getBuildStatusByBuildKeyId((int) $build_key_id);
 
-        \App\Build::updateBuildStatus($build_key_id, $status);
+        Build::updateBuildStatus((int) $build_key_id, $status);
     }
 }
