+++ app/Console/PCITDaemon/Commands/MigrateCommand.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Console\PCITDaemon;
+namespace App\Console\PCITDaemon\Commands;
 
+use App\Console\PCITDaemon\Migrate;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
