+++ bin/src/pcitd.php
@@ -2,9 +2,10 @@
 
 <?php
 
-use App\Console\PCITDaemon\AgentCommand;
-use App\Console\PCITDaemon\MigrateCommand;
-use App\Console\PCITDaemon\UpCommand;
+use App\Console\PCITDaemon\Commands\AgentCommand;
+use App\Console\PCITDaemon\Commands\MigrateCommand;
+use App\Console\PCITDaemon\Commands\ServerCommand;
+use App\Console\PCITDaemon\Commands\UpCommand;
 use Symfony\Component\Console\Application;
 
 require __DIR__.'/../../framework/bootstrap/app.php';
@@ -19,6 +20,8 @@ try {
     $cli->add(new MigrateCommand());
 
     $cli->add(new UpCommand());
+
+    $cli->add(new ServerCommand());
 
     $cli->add(new AgentCommand());
